feat(aitrader): Phase 2 — add proper AI Analyse and Watch Trades controllers with spec-compliant endpoints
