+++ src/main/java/com/dtech/aitrader/web/AiLevelsController.java
@@ -1,12 +1,9 @@
 package com.dtech.aitrader.web;
 
 import com.dtech.aitrader.data.AiLevels;
-import com.dtech.aitrader.data.WatchTrade;
 import com.dtech.aitrader.repository.AiLevelsRepository;
-import com.dtech.aitrader.repository.WatchTradeRepository;
 import com.dtech.aitrader.service.AiDrawingsMerger;
 import com.dtech.aitrader.service.AiBatchRunnerService;
-import com.dtech.aitrader.service.AiAnalyseService;
 import com.dtech.aitrader.service.LevelsAgentService;
 import com.dtech.aitrader.service.MergeResult;
 import com.dtech.kitecon.repository.IndexSymbolRepository;
@@ -17,7 +14,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
-import lombok.Builder;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.*;
@@ -29,11 +25,9 @@ import java.util.stream.Collectors;
 @Slf4j
 public class AiLevelsController {
     private final LevelsAgentService levelsAgentService;
-    private final AiAnalyseService aiAnalyseService;
     private final AiDrawingsMerger aiDrawingsMerger;
     private final AiBatchRunnerService aiBatchRunnerService;
     private final AiLevelsRepository aiLevelsRepository;
-    private final WatchTradeRepository watchTradeRepository;
     private final IndexSymbolRepository indexSymbolRepository;
     private final ObjectMapper objectMapper;
 
@@ -195,41 +189,6 @@ public class AiLevelsController {
         }
     }
 
-    @PostMapping("/analyse/run")
-    public AiAnalyseResponse analyseRun(
-            @RequestBody AiAnalyseRequest request,
-            Authentication auth) {
-
-        log.info("POST /api/ai-levels/analyse/run - symbol={}, tabId={}, layoutId={}, timeframe={}, user={}",
-                request.getSymbol(), request.getTabId(), request.getLayoutId(), request.getTimeframe(), auth.getName());
-
-        try {
-            Long userId = extractUserId(auth);
-
-            // Run AI Analyse service
-            List<WatchTrade> watchTrades = aiAnalyseService.runForSymbol(
-                    request.getSymbol(),
-                    request.getTabId(),
-                    request.getLayoutId(),
-                    request.getTimeframe(),
-                    userId
-            );
-
-            // Build response
-            AiAnalyseResponse response = AiAnalyseResponse.builder()
-                    .symbol(request.getSymbol())
-                    .watchTrades(watchTrades)
-                    .generatedAt(LocalDate.now().toString())
-                    .build();
-
-            return response;
-
-        } catch (Exception e) {
-            log.error("Error in /api/ai-levels/analyse/run", e);
-            throw new RuntimeException("AI Analyse run failed: " + e.getMessage(), e);
-        }
-    }
-
     private Long extractUserId(Authentication auth) {
         if (auth == null || auth.getName() == null) {
             throw new IllegalStateException("User not authenticated");
@@ -237,38 +196,4 @@ public class AiLevelsController {
         // Return anand's user ID (hardcoded for now; will be fixed with proper UserRepository integration)
         return 1L;
     }
-}
-
-// Request/Response DTOs
-class AiAnalyseRequest {
-    public String symbol;
-    public String tabId;
-    public Long layoutId;
-    public String timeframe;
-
-    public String getSymbol() { return symbol; }
-    public void setSymbol(String symbol) { this.symbol = symbol; }
-    public String getTabId() { return tabId; }
-    public void setTabId(String tabId) { this.tabId = tabId; }
-    public Long getLayoutId() { return layoutId; }
-    public void setLayoutId(Long layoutId) { this.layoutId = layoutId; }
-    public String getTimeframe() { return timeframe; }
-    public void setTimeframe(String timeframe) { this.timeframe = timeframe; }
-}
-
-class AiAnalyseResponse {
-    public String symbol;
-    public List<WatchTrade> watchTrades;
-    public String generatedAt;
-
-    @lombok.Builder
-    public AiAnalyseResponse(String symbol, List<WatchTrade> watchTrades, String generatedAt) {
-        this.symbol = symbol;
-        this.watchTrades = watchTrades;
-        this.generatedAt = generatedAt;
-    }
-
-    public String getSymbol() { return symbol; }
-    public List<WatchTrade> getWatchTrades() { return watchTrades; }
-    public String getGeneratedAt() { return generatedAt; }
 }
